refactor(spec): split operation_metadata into focused v4 modules
The module had grown past 1,000 lines mixing the artifact model, the
validated plan, structural IR validation, policy validation, and
projection reconciliation. Split it into operation_metadata/{model,
plan, structural, policy} and move projection input synchronization to
projections/sync alongside the other projection code. Public API is
unchanged via re-exports.

Addresses cubic review on #1884 (operation_metadata.rs:767).
