+++ crates/coral-spec/src/v4/projections/mod.rs
@@ -3,6 +3,7 @@ mod model;
 mod names;
 mod pagination;
 mod runtime;
+mod sync;
 
 pub use derive::generate_projection_catalog;
 pub use model::*;
@@ -11,3 +12,4 @@ pub use runtime::{
     mcp_projection_arg_specs, projection_arg_specs, projection_column_specs,
     projection_filter_specs, request_spec_for_projection,
 };
+pub use sync::{ProjectionInputSyncMode, sync_projection_inputs};
